Fixing attachment replacing.
Theoretically, the range for previously attachment shouldn't change, as we are appending new text lineary.
But in fact it's the same or differs by 1 index.
The proper solution would be to save attachements and after layouting whole text, search for them and their proper ranges.
But due to lack of time I implemented check for nearest neighbour and check which one is @c NSAttachment.